+++ tests/playwright/sanity/modules/v4-tests/editor-variables/utils.ts
@@ -1,8 +1,30 @@
 import { type Page } from '@playwright/test';
 import EditorSelectors from '../../../../selectors/editor-selectors';
+import WpAdminPage from '../../../../pages/wp-admin-page';
+
+export const initTemplate = async ( page: Page ) => {
+	const adminPage = new WpAdminPage( page );
+	await adminPage.setExperiments( { e_variables_manager: 'active' } );
+	const editorPage = await adminPage.openNewPage( false, false );
+	await editorPage.loadTemplate( 'tests/playwright/templates/default-v4.json' );
+};
+
+export const addFontVariable = async ( page: Page ) => {
+	const variableData = { name: `test-font-variable-${ Date.now() }`, value: 'Arial', type: 'font' as const };
+	await addVariable( page, variableData, 'Typography', 'font-family' );
+	await detachVariable( page, 'Typography', 'Font Family', variableData.name );
+	return variableData;
+};
+
+export const addColorVariable = async ( page: Page ) => {
+	const variableData = { name: `test-color-variable-${ Date.now() }`, value: '#ff0000', type: 'color' as const };
+	await addVariable( page, variableData, 'Typography', 'text-color' );
+	await detachVariable( page, 'Typography', 'Text Color', variableData.name );
+	return variableData;
+};
 
 export const getControl = async ( page: Page, styleSectionSelector: string, controlToAccessFrom: string ) => {
-	await page.frameLocator( EditorSelectors.canvas ).getByText( 'This is a title' ).click();
+	await page.frameLocator( EditorSelectors.canvas ).locator( 'body' ).getByText( 'This is a title' ).click();
 	await page.getByRole( 'tab', { name: 'Style' } ).click();
 	if ( await page.locator( `#${ controlToAccessFrom }-control` ).isHidden() ) {
 		await page.getByRole( 'button', { name: styleSectionSelector } ).click();
@@ -34,7 +56,7 @@ export const addVariable = async ( page: Page, variable: { name: string, value: 
 };
 
 export const detachVariable = async ( page: Page, styleSectionSelector: string, controlLabelToDetach: string, variableName: string ) => {
-	await page.frameLocator( EditorSelectors.canvas ).getByText( 'This is a title' ).click();
+	await page.frameLocator( EditorSelectors.canvas ).locator( 'body' ).getByText( 'This is a title' ).click();
 	await page.getByRole( 'tab', { name: 'Style' } ).click();
 	if ( await page.getByText( controlLabelToDetach ).isHidden() ) {
 		await page.getByRole( 'button', { name: styleSectionSelector } ).click();
